测试新版图像代码并修复部分问题，能成功跑无元素方框

+++ Macro_TC264/code/image_process.h
@@ -45,9 +45,9 @@ typedef struct
 
 typedef struct
 {
-    int left;
-    int right;
-    int T;
+    uint8 left;
+    uint8 right;
+    uint8 T;
 }corner_result;
